Move 3ds services in example to 'service.threeds'
