+++ example/src/main/kotlin/com/processout/example/service/threeds/Checkout3DSServiceDelegate.kt
@@ -1,4 +1,4 @@
-package com.processout.example.service
+package com.processout.example.service.threeds
 
 import android.app.Activity
 import android.net.Uri
