+++ example/src/main/kotlin/com/processout/example/service/threeds/POAdyen3DSService.kt
@@ -1,4 +1,4 @@
-package com.processout.example.service
+package com.processout.example.service.threeds
 
 import android.app.Activity
 import com.adyen.threeds2.*
